requirements.txt
streamlit
google-genai
sentence-transformers
faiss-cpu
numpy